Migrate Video class to "core" subpackage
