+++ youtubesearchpython/__init__.py
@@ -1,7 +1,7 @@
 from youtubesearchpython.search import Search, VideosSearch, ChannelsSearch, PlaylistsSearch, CustomSearch, ChannelSearch
 from youtubesearchpython.extras import Video, Playlist, Suggestions, Hashtag
 from youtubesearchpython.streamurlfetcher import StreamURLFetcher
-from youtubesearchpython.internal.constants import *
+from youtubesearchpython.core.constants import *
 
 
 __title__        = 'youtube-search-python'
